+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## v3.4.0
+
+### Added
+
+- Added vehicle model types specifications `types.json`
+- **PHP SDK** method `getVehicleModelsTypesSpecification`
+- **PHP SDK** protected method `getVehicleTypeAliasById`
+- **PHP SDK** protected method `getVehicleModelsSpecificationFilePath`
+- Extended support for models of cars, new model files are grouped by vehicle type (example: `models_bus.json` - cars)
+
 ## v3.3.1
 
 ### Fixed